Check the return value of mkdirs() and handle errors. Issue: RAT-109 Contributed By: Hugo Hirsch
git-svn-id: https://svn.apache.org/repos/asf/incubator/rat/main/trunk@1208284 13f79535-47bb-0310-9956-ffa450edef68

+++ apache-rat-tasks/src/test/java/org/apache/rat/anttasks/ReportTest.java
@@ -40,7 +40,9 @@ public class ReportTest extends AbstractRatAntTaskTest {
 
     public void testWithReportSentToFile() throws Exception {
         final File reportFile = new File(getTempDir(), "selftest.report");
-        getTempDir().mkdirs();
+        if(!getTempDir().mkdirs()) {
+            throw new IOException("Could not create temporary directory " + getTempDir());
+        }
         final String alLine = "AL +\\Q" + getAntFileName() + "\\E";
         if (reportFile.isFile()  &&  !reportFile.delete()) {
             throw new IOException("Unable to remove report file " + reportFile);
